Update 1.2
+Admin panel for role and permission
+Middlewares
+Export to PDF and  Excel

+++ routes/web.php
@@ -3,6 +3,7 @@
 use Illuminate\Support\Facades\Route;
 use App\Http\Models\Flower;
 use App\Http\Controllers\FlowerController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\DB;
 
 /*
@@ -27,14 +28,23 @@ Route::get('/dashboard', function () {
 })->middleware(['auth'])->name('dashboard');
 
 
-Route::get('/flower/cards', [FlowerController::class, 'cardView']);
-Route::get('/flower/filter', [FlowerController::class, 'filter']);
-Route::get('/flower/create', [FlowerController::class, 'create']);
-Route::post('/flower', [FlowerController::class, 'store']);
-Route::get('/flower/{flowers}', [FlowerController::class, 'show']);
-Route::get('/flower/{flowers}/edit', [FlowerController::class, 'edit']);
-Route::put('/flower/{flowers}', [FlowerController::class, 'update']);
-Route::delete('/flower/{flowers}', [FlowerController::class, 'destroy']);
+//Flower Routes
+Route::get('/flower/print', [FlowerController::class, 'exportPdf'])->middleware(['auth']);
+Route::get('/flower/excel', [FlowerController::class, 'exportExcel'])->middleware(['auth']);
+Route::get('/flower/cards', [FlowerController::class, 'cardView'])->middleware(['auth']);
+Route::get('/flower/filter', [FlowerController::class, 'filter'])->middleware(['auth']);
+Route::get('/flower/create', [FlowerController::class, 'create'])->middleware(['auth']);
+Route::post('/flower', [FlowerController::class, 'store'])->middleware(['auth']);
+Route::get('/flower/{flowers}', [FlowerController::class, 'show'])->middleware(['auth']);
+Route::get('/flower/{flowers}/edit', [FlowerController::class, 'edit'])->middleware(['auth']);
+Route::get('/flower/{flowers}/change', [FlowerController::class, 'change'])->middleware(['auth']);
+Route::put('/flower/{flowers}/upload', [FlowerController::class, 'upload'])->middleware(['auth']);
+Route::put('/flower/{flowers}', [FlowerController::class, 'update'])->middleware(['auth']);
+Route::delete('/flower/{flowers}', [FlowerController::class, 'destroy'])->middleware(['auth']);
 
+//User Routes
+Route::get('/user/create', [UserController::class, 'create'])
+->middleware(['auth','permission:users-create']);
+Route::post('/user', [UserController::class, 'store'])->middleware(['auth']);
 
 require __DIR__.'/auth.php';
